# Conflicts:
#	src/main/java/vn/edu/fpt/petworldplatform/config/SecurityConfig.java
#	src/main/java/vn/edu/fpt/petworldplatform/controller/AuthController.java
#	src/main/java/vn/edu/fpt/petworldplatform/controller/CustomerController.java

+++ src/main/java/vn/edu/fpt/petworldplatform/repository/StaffRepo.java
@@ -7,9 +7,7 @@ import vn.edu.fpt.petworldplatform.entity.Staff;
 import java.util.Optional;
 
 @Repository
-public interface StaffRepo extends JpaRepository<Staff, Integer> {
+public interface StaffRepo extends JpaRepository<Staff, Long> {
     Optional<Staff> findByUsername(String username);
-
     Optional<Staff> findByEmail(String email);
-
 }
